fix: options passed to useField will not always cause a re-render
